rename api to dpg_butler
